Added new styles and changed styles
