+++ public/styles.css
@@ -1,39 +1,53 @@
 /* styles.css */
+body {
+  background-color: lightgrey;
+  font-family: Arial, sans-serif;
+  margin: 0, auto;
+  display: flex;
+  text-align: center;
+  justify-content: center;
+}
+
 .main-container {
-  background-color: #f0f0f0;
-  padding: 20px;
+  background-color: whitesmoke;
+  /* padding: 20px; */
   border-radius: 10px;
-  width: 300px;
+  width: fit-content;
   margin: 0 auto;
-  box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.1);
   text-align: left;
   display: flex;
   flex-direction: column;
 }
 
 h1 {
-  color: #333;
-  font-size: 1rem;
+  color: gray;
+  text-wrap: nowrap;
 }
 
 input[type="checkbox"] {
-  width: 20px;
-  height: 20px;
+  width: 30px;
+  height: 30px;
+  margin-left: 3px;
 }
 
 .card {
   margin: 10px;
   padding: 10px;
   border-radius: 5px;
-  box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.1);
+  box-shadow: 0px 4px 6px rgba(214, 89, 89, 0.541);
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  font-size: 0.35rem;
 }
 
 .card-buttons {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: flex-start;
+  display: inline-block;
+  flex-wrap: nowrap;
+  justify-content: flex-end;
 }
 
 .card-button {
+  font-size: 8px;
   margin: 5px;
 }
