+++ src/App.jsx
@@ -1,4 +1,3 @@
-import "./App.css";
 import Header from "./components/Header";
 import MainC from "./components/MainC";
 import Footer from "./components/Footer";
